Rename CLI implementation files to make codecov happy.
