+++ cfg_parser.py
@@ -3,7 +3,7 @@
 #
 # Copyright (c) 2017-2021 OWS Contributors
 # SPDX-License-Identifier: Apache-2.0
-from datacube_ows.cfg_parser import main
+from datacube_ows.cfg_parser_impl import main
 from datacube_ows.startup_utils import initialise_debugging
 
 if __name__ == '__main__':
