+++ update_ranges.py
@@ -3,7 +3,7 @@
 #
 # Copyright (c) 2017-2021 OWS Contributors
 # SPDX-License-Identifier: Apache-2.0
-from datacube_ows.update_ranges import main
+from datacube_ows.update_ranges_impl import main
 
 if __name__ == '__main__':
     main()
